introduce stream dead policy
